拆分星火燎原，按时间顺序调整武将包顺序
