+++ character/sp/voices.js
@@ -795,4 +795,62 @@ export default {
 	"#olqingya1": "罡风从虎，威震四方！",
 	"#olqingya2": "铁车过处，寸草不生！",
 	"#yadan:die": "多谢丞相不杀之恩……",
+	"#xinfu_sanwen1": "文若春华，思若泉涌。",
+	"#xinfu_sanwen2": "独步汉南，散文天下。",
+	"#xinfu_qiai1": "未知身死处，何能两相完？",
+	"#xinfu_qiai2": "悟彼下泉人，喟然伤心肝。",
+	"#xinfu_denglou1": "登兹楼以四望兮，聊暇日以销忧。",
+	"#xinfu_denglou2": "惟日月之逾迈兮，俟河清其未极。",
+	"#wangcan:die": "一作驴鸣悲，万古送葬别……",
+	"#xinfu_jixu1": "击虚箭射，懈敌戒备。",
+	"#xinfu_jixu2": "虚实难辨，方迷敌方之心！",
+	"#sp_taishici:die": "刘繇之见，短浅也……",
+	"#xinfu_guolun1": "品过是非，讨评好坏。",
+	"#xinfu_guolun2": "若有天下太平时，必讨四海之内才。",
+	"#xinfu_songsang1": "送丧至东吴，使命已完。",
+	"#xinfu_songsang2": "送丧虽至，吾与孝则得相交。",
+	"#re_jsp_pangtong:die": "我终究……不得东吴赏识……",
+	"#qinguo_use1": "为国勤事，体素精勤。",
+	"#qinguo_use2": "忠勤为国，通达治体。",
+	"#lvdai:die": "再也不能，为吴国奉身了……",
+	"#xinfu_jijun1": "集万千义军，定天下大局！",
+	"#xinfu_jijun2": "集民力万千，亦可为军！",
+	"#xinfu_fangtong1": "统领方队，为民意所举！",
+	"#xinfu_fangtong2": "三十六方，必为大统！",
+	"#re_zhangliang:die": "人公也难逃被人所杀……",
+	"#xinfu_weilu1": "贼人势大，需从长计议。",
+	"#xinfu_weilu2": "时机未到，先行撤退。",
+	"#xinfu_zengdao1": "有功赏之，有过罚之。",
+	"#xinfu_zengdao2": "治军之道，功过分明。",
+	"#lvqian:die": "我自泰山郡以来，百姓获安，镇军伐贼，此生已无憾！",
+	"#xinfu_guanwei1": "今日宴请诸位，有要事相商。",
+	"#xinfu_guanwei2": "天下未定，请主公以大局为重。",
+	"#xinfu_gongqing1": "尔辈何故与降虏交善。",
+	"#xinfu_gongqing2": "豪将在外，增兵必成祸患啊！",
+	"#panjun:die": "耻失荆州，耻失荆州啊！",
+	"#xinfu_andong1": "勇足以当大难，智涌以安万变。",
+	"#xinfu_andong2": "宽猛克济，方安河东之民。",
+	"#xinfu_yingshi1": "应民之声，势民之根。",
+	"#xinfu_yingshi2": "应势而谋，顺民而为。",
+	"#duji:die": "试船而溺之，虽亡而忠至……",
+	"#xinfu_duanfa1": "身体发肤，受之父母。",
+	"#xinfu_duanfa2": "今断发以明志，尚不可证吾之心意？",
+	"#xinfu_youdi1": "东吴已容不下我，愿降以保周全。",
+	"#xinfu_youdi2": "笺书七条，足以表我归降之心。",
+	"#zhoufang:die": "功亏一篑，功亏一篑啊……",
+	"#xinfu_guanchao1": "朝夕之间，可知所进退。",
+	"#xinfu_guanchao2": "月盈，潮起晨暮也；月亏，潮起日半也。",
+	"#xinfu_xunxian1": "督军之才，子明强于我甚多。",
+	"#xinfu_xunxian2": "此间重任，公卿可担之。",
+	"#yanjun:die": "著作，还……没完成……",
+	"#xinfu_kannan1": "俊才之杰，材匪戡难。",
+	"#xinfu_kannan2": "戡，克也，难，攻之。",
+	"#liuyao:die": "伯符小儿，还我子义！",
+	"#xinfu_tushe1": "非英杰不图？吾既谋之且射毕！",
+	"#xinfu_tushe2": "汉室衰微，朝纲祸乱，必图后福。",
+	"#xinfu_limu1": "米贼作乱，吾必为益州自保。",
+	"#xinfu_limu2": "废史立牧，可得一方安定。",
+	"#liuyan:die": "背疮难治，世子难继……",
+	"#xinfu_zhanji1": "公瑾安全至吴，心安之。",
+	"#xinfu_zhanji2": "功曹之恩，吾必有展骥之机。",
 };
