--- character/xinghuoliaoyuan/card.js
@@ -1,4 +1,0 @@
-import { lib, game, ui, get, ai, _status } from "../../noname.js";
-
-const cards = {};
-export default cards;
