--- character/xinghuoliaoyuan/characterFilter.js
@@ -1,5 +1,0 @@
-import { lib, game, ui, get, ai, _status } from "../../noname.js";
-
-const characterFilters = {};
-
-export default characterFilters;
