--- character/xinghuoliaoyuan/pinyin.js
@@ -1,2 +1,0 @@
-const pinyins = {};
-export default pinyins;
