--- character/xinghuoliaoyuan/sort.js
@@ -1,19 +1,0 @@
-const characterSort = {
-	// xinghuoliaoyuan_tianfu:[],
-	xinghuoliaoyuan_tianliang: ["duji", "liuyan", "yanjun"],
-	xinghuoliaoyuan_tianji: ["panjun", "wangcan"],
-	xinghuoliaoyuan_tiantong: ["re_jsp_pangtong", "sp_taishici"],
-	xinghuoliaoyuan_tianxiang: ["lvdai", "zhoufang", "liuyao"],
-	xinghuoliaoyuan_qisha: ["lvqian", "re_zhangliang"],
-};
-
-const characterSortTranslate = {
-	xinghuoliaoyuan_tianfu: "天府",
-	xinghuoliaoyuan_tianliang: "天梁",
-	xinghuoliaoyuan_tianji: "天机",
-	xinghuoliaoyuan_tiantong: "天同",
-	xinghuoliaoyuan_tianxiang: "天相",
-	xinghuoliaoyuan_qisha: "七杀",
-};
-
-export { characterSort, characterSortTranslate };
