+++ game/package.js
@@ -1,28 +1,36 @@
 window.noname_package = {
 	character: {
+		//经典标包＋神话再临
 		standard: "标准",
-		refresh: "界限突破",
 		shenhua: "神话再临",
+		extra: "神将",
+		//一将成名+界限突破
 		yijiang: "一将成名",
-		newjiang: "新一将成名",
-		sixiang: "四象封印",
+		refresh: "界限突破",
+		//新时代通用武将
 		sp: "璀璨星河",
+		sp2: "系列专属",
+		newjiang: "新一将成名",
+		//OL服
 		onlyOL: "OL专属",
 		yingbian: "文德武备",
 		clan: "门阀士族",
-		xinghuoliaoyuan: "星火燎原",
+		//十周年服
 		huicui: "群英荟萃",
 		xianding: "限定专属",
-		sp2: "系列专属",
-		extra: "神将",
+		//移动服+外服
 		mobile: "移动版",
 		shiji: "始计篇",
 		sb: "谋攻篇",
 		tw: "外服武将",
+		//各服联动+怀旧
 		collab: "联动卡",
+		old: "怀旧",
+		//线下系列
 		offline: "线下武将",
 		jsrg: "江山如故",
-		old: "怀旧",
+		sixiang: "四象封印",
+		//DIY系列
 		diy: "DIY",
 		ddd: "3D精选",
 		key: "Key",
