+++ game/source.js
@@ -375,19 +375,6 @@ window.noname_source_list = [
 	"character/xianding/sort.js",
 	"character/xianding/translate.js",
 	"character/xianding/voices.js",
-	"character/xinghuoliaoyuan/card.js",
-	"character/xinghuoliaoyuan/character.js",
-	"character/xinghuoliaoyuan/characterFilter.js",
-	"character/xinghuoliaoyuan/characterReplace.js",
-	"character/xinghuoliaoyuan/dynamicTranslate.js",
-	"character/xinghuoliaoyuan/index.js",
-	"character/xinghuoliaoyuan/intro.js",
-	"character/xinghuoliaoyuan/perfectPairs.js",
-	"character/xinghuoliaoyuan/pinyin.js",
-	"character/xinghuoliaoyuan/skill.js",
-	"character/xinghuoliaoyuan/sort.js",
-	"character/xinghuoliaoyuan/translate.js",
-	"character/xinghuoliaoyuan/voices.js",
 	"character/yijiang/card.js",
 	"character/yijiang/character.js",
 	"character/yijiang/characterFilter.js",
